fixed navbar css and added confirm message whene deleting a post

+++ frontend/src/components/posts/Comments/Comments.jsx
@@ -35,7 +35,10 @@ export default function Comments({ post, user }) {
   };
 
   const deleteComment = (comment) => {
-    if (comment.user.email === user.email) {
+    if (
+      comment.user.email === user.email &&
+      window.confirm("Delete the item?")
+    ) {
       Api.delete("/comments/" + comment.id).then((r) => getAllByPost());
     } else {
       window.alert("Only the user who created the comment can delete it");
